+++ source/test/cpp/test_superalloc.cpp
@@ -162,7 +162,7 @@ UNITTEST_SUITE_BEGIN(main_allocator)
                     void** store     = (void**)(allocdmp + (offset * 8));
                     void*  alloc_ptr = *store;
                     u32    ptr_size  = s_alloc.get_size(alloc_ptr);
-                    CHECK_TRUE(size <= ptr_size);
+                    CHECK_TRUE((u32)size <= ptr_size);
                     u32 ptr_tag = s_alloc.get_tag(alloc_ptr);
                     CHECK_EQUAL(size, ptr_tag);
                     s_alloc.deallocate(alloc_ptr);
